Toddbell/edex anywhere (#102)
* Minimal error fixes for moving editor extensions

generalizing the search folder for our GUISkin. This is the only exception that does pop up when we move the editor extensions to a different folder. Most other issues with the SDK have work arounds (found sdk in a different directory? drag and drop that onto the panel so we can still call it wherever they wanted it to).

Also simplifies some LastIndexOf for Assets where if we look for a "/" sometimes will return a -1 on the path (because we may not know if we will be given a \\ instead which will not be detected and throw another error).

* Update README.md

Adding some documentation regarding moving the editor extension to anywhere but the /Assets/ root folder.

* Update PlayFabEditorPrefsSO.cs

Added a TODO to remove hardcoding the Assets path.

* Update PlayFabEditorPrefsSO.cs

quick fix