Toddbell/asmdef fix (#112)
* Adding .asmdef files to the extensions repo

this should prevent the Unity editor from re-compiling our code every time a developer alt-tab's into the editor.

* changed editor extensions asmdef to a simplified version

We basically don't care about target platforms for EditorExtensions since they should only run in the Editor. We will however need to add to SDKGenerator to add an asmdef to the Unity SDK.

* Delete PlayFabEditorExtensions.asmdef

so that made the editor stop rendering :(

* Create PlayFabEditorExtensions.asmdef

This should prevent the editor extensions from re-compiling every time a user navigates back to Unity. This file is expected to exist for Unity packages.